Add docs and adjust flow + babel build
