+++ package.json
@@ -105,6 +105,7 @@
     "babel-loader": "7.0.0",
     "babel-plugin-inline-svg": "0.1.0",
     "babel-plugin-transform-class-properties": "6.24.1",
+    "babel-plugin-transform-flow-strip-types": "6.22.0",
     "babel-preset-env": "1.6.1",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-flow": "6.23.0",
