+++ src/components/Pop/README.md
@@ -1,0 +1,14 @@
+# Pop
+
+A Pop component is a Component that abstracts logic related to mounting and auto-positioning the child component to the DOM, relative to it's original target element.
+
+This component is powered by [Popper.js](https://popper.js.org/) the evolution of [Drop](../Drop).
+
+## Components
+
+The Toolbar component is comprised of smaller components:
+
+* [Arrow](./docs/Arrow.md)
+* [Manager](./docs/Manager.md)
+* [Popper](./docs/Popper.md)
+* [Reference](./docs/Reference.md)
