+++ src/components/Pop/docs/Manager.md
@@ -1,0 +1,3 @@
+# Manager
+
+A Pop.Manager is a context provider for the child [Popper](./Popper.md) and [Reference](./Reference.md) components. This pattern is part of [Popper](../Popper).
