+++ src/components/Pop/docs/Reference.md
@@ -1,0 +1,14 @@
+# Reference
+
+A Pop.Reference is a tiny wrapper that renders the trigger UI, powered by [Popper.js](https://popper.js.org/).
+
+This component is powered by [Popper.js](https://popper.js.org/) the evolution of [Drop](../Drop).
+
+## Example
+
+```jsx
+<Pop>
+  ...
+  <Pop.Reference>Hover me!</Pop.Reference>
+</Pop>
+```
