+++ src/components/Tooltip/docs/Tooltip.md
@@ -1,0 +1,19 @@
+# Toolbar
+
+A Toolbar component is UI component that provides the user with additional context when engaged.
+
+## Example
+
+```jsx
+<Tooltip title="You're my boy!!!">Blue</Tooltip>
+```
+
+## Props
+
+| Prop          | Type       | Description                                                         |
+| ------------- | ---------- | ------------------------------------------------------------------- |
+| className     | `string`   | Custom class names to be added to the component.                    |
+| renderContent | `function` | Renders a component within the Tooltip. Is prioritized over `title` |
+| title         | `string`   | Text to display within the Tooltip.                                 |
+
+This component is powered by [Pop](../Pop). After additional props, check out [Pop's documentation](../Pop).
