--- src/components/Tooltip/types.js
@@ -1,6 +1,0 @@
-// @flow
-import type { Props as PopTypes } from '../Pop/types'
-
-export type Props = {
-  ...$Exact<PopTypes>,
-}
